requirements.txt
pandas
requests
matplotlib
seaborn
python-dotenv
ipykernel
tabulate
numpy
logging